Move Python API code to separate file
