+++ src/Drivers/Common.items
@@ -232,6 +232,8 @@ by the PreBuild.targets file.
     <Environments Include="$(SolutionDir)Environments\Xbox\bin\$(StandardOutputDir)\Reko.Environments.Xbox.*" />
     <Environments Include="$(SolutionDir)Environments\ZX81\bin\$(StandardOutputDir)\Reko.Environments.ZX81.*" />
 
+    <Scripts Include="$(SolutionDir)Scripts\Python\api.py" />
+
     <SymbolSources Include="$(SolutionDir)Sources\LGSymLoader\bin\$(Configuration)\Reko.Symbols.LGSymLoader.*" />
 
     <Libraries Include="$(SolutionDir)Libraries\Libc\bin\$(StandardOutputDir)\Reko.Libraries.Libc.*" />
