+++ src/Scripts/Python/PythonAPI.cs
@@ -19,6 +19,8 @@
 #endregion
 
 using Microsoft.Scripting.Hosting;
+using Reko.Core;
+using Reko.Core.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -33,9 +35,9 @@ namespace Reko.Scripts.Python
     {
         private readonly ScriptScope scope;
 
-        public PythonAPI(ScriptEngine engine)
+        public PythonAPI(IServiceProvider services, ScriptEngine engine)
         {
-            scope = EvaluatePythonDefinitions(engine);
+            scope = EvaluatePythonDefinitions(services, engine);
         }
 
         public object CreateProgramWrapper(RekoAPI rekoAPI)
@@ -45,101 +47,13 @@ namespace Reko.Scripts.Python
             return programCreator(rekoAPI);
         }
 
-        private ScriptScope EvaluatePythonDefinitions(ScriptEngine engine)
+        private ScriptScope EvaluatePythonDefinitions(
+            IServiceProvider services, ScriptEngine engine)
         {
-            var pythonAPIDefinitions = @"
-class Comments(object):
-    def __init__(self, reko):
-        self._reko = reko
-
-    def __setitem__(self, addr, comment):
-        if not isinstance(comment, str):
-            raise TypeError(
-                'Unsupported type: {}. Expected type: str'.format(
-                    type(comment)))
-        self._reko.SetUserComment(addr, comment)
-
-class MemorySlice(object):
-    def __init__(self, reko, start_addr):
-        self._reko = reko
-        self._start_addr = start_addr
-
-    @property
-    def int32(self):
-        return self._reko.ReadInt32(self._start_addr)
-
-    @property
-    def c_str(self):
-        return self._reko.ReadCString(self._start_addr)
-
-class Memory(object):
-    def __init__(self, reko):
-        self._reko = reko
-
-    def __getitem__(self, addr):
-        return MemorySlice(self._reko, addr)
-
-class Globals(object):
-    def __init__(self, reko):
-        self._reko = reko
-
-    def __setitem__(self, addr, decl):
-        if isinstance(decl, str):
-            self._reko.SetUserGlobal(addr, decl)
-            return
-        raise TypeError('Unsupported type: {}'.format(type(decl)))
-
-class Procedure(object):
-    def __init__(self, reko, addr):
-        self._reko = reko
-        self._addr = addr
-
-    @property
-    def decompile(self):
-        raise AttributeError('Can not read decompile attribute')
-
-    @decompile.setter
-    def decompile(self, value):
-        self._reko.SetUserProcedureDecompileFlag(self._addr, value)
-
-class Procedures(object):
-    def __init__(self, reko):
-        self._reko = reko
-
-    def __getitem__(self, addr):
-        return Procedure(self._reko, addr)
-
-    def __setitem__(self, addr, proc):
-        if isinstance(proc, str):
-            self._reko.SetUserProcedure(addr, proc)
-            return
-        raise TypeError('Unsupported type: {}'.format(type(proc)))
-
-class Program(object):
-    def __init__(self, reko):
-        self._comments = Comments(reko)
-        self._memory = Memory(reko)
-        self._globals = Globals(reko)
-        self._procedures = Procedures(reko)
-
-    @property
-    def comments(self):
-        return self._comments
-
-    @property
-    def memory(self):
-        return self._memory
-
-    @property
-    def globals(self):
-        return self._globals
-
-    @property
-    def procedures(self):
-        return self._procedures
-";
             var scope = engine.CreateScope();
-            engine.Execute(pythonAPIDefinitions, scope);
+            var cfgSvc = services.RequireService<IConfigurationService>();
+            var pythonAPIFile = cfgSvc.GetInstallationRelativePath("api.py");
+            engine.ExecuteFile(pythonAPIFile, scope);
             return scope;
         }
     }
